refactor(#741): refactor geolocController
