+++ api/controllers/v1/geoloc/count-entrances.js
@@ -1,0 +1,20 @@
+const ErrorService = require('../../../services/ErrorService');
+const GeoLocService = require('../../../services/GeoLocService');
+
+module.exports = async (req, res) => {
+  const { southWestBound, northEastBound, errorMessage } =
+    GeoLocService.checkAndGetCoordinatesParams(req);
+
+  if (errorMessage !== '') return res.badRequest(errorMessage);
+
+  try {
+    const result = await GeoLocService.countEntrances(
+      southWestBound,
+      northEastBound
+    );
+    return res.json({ count: result });
+  } catch (e) {
+    ErrorService.getDefaultErrorHandler(res)(e);
+    return false;
+  }
+};
